+++ melite/__init__.py
@@ -6,8 +6,7 @@ Public API
 The following symbols are part of the stable public API:
 
     from melite import Config
-    from melite import load_dataset
-    from melite import ResultManager
+    from melite import load_datasets
     from melite import plot_cv_distributions
     from melite import predict
     from melite import __version__
@@ -16,16 +15,14 @@ The following symbols are part of the stable public API:
 import logging
 
 from .config import Config
-from .load_dataset import load_dataset
-from .result_manager import ResultManager
+from .load_dataset import load_datasets
 from .plot_metrics import plot_cv_distributions
 from .predict import predict
 from .version import __version__
 
 __all__ = [
     "Config",
-    "load_dataset",
-    "ResultManager",
+    "load_datasets",
     "plot_cv_distributions",
     "predict",
     "__version__",